main_service/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioredis==2.0.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncio==4.0.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
disutils==1.4.32.post2
fastapi==0.115.14
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.2.0
requests==2.32.5
rq==2.4.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==3.0.2
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
yarl==1.20.1
